pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bssync"
version = "0.1.0"
description = "Two-way sync between local markdown files and a BookStack wiki"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Alex Rybkin" }]
requires-python = ">=3.9"
keywords = ["bookstack", "wiki", "markdown", "sync", "documentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
build = [
    "pyinstaller>=6.0",
]

[project.scripts]
bssync = "bssync.cli:main"

[project.urls]
Homepage = "https://github.com/alexamber/bssync"
Issues = "https://github.com/alexamber/bssync/issues"
Repository = "https://github.com/alexamber/bssync"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
